requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1 